+++ src/sparse/matvec.rs
@@ -37,9 +37,9 @@ pub fn matvec_csr<T: Scalar>(m: &CsrMatrix<T>, x: &[T]) -> Result<Vec<T>, Dimens
     let col_idx = m.col_indices();
     let vals = m.values();
     for r in 0..m.rows() {
-        for k in row_ptr[r]..row_ptr[r + 1] {
+        for k in row_ptr[r] as usize..row_ptr[r + 1] as usize {
             let prev = out[r];
-            out[r] = prev.add(vals[k].mul(x[col_idx[k]]));
+            out[r] = prev.add(vals[k].mul(x[col_idx[k] as usize]));
         }
     }
     Ok(out)
@@ -75,8 +75,8 @@ pub fn matvec_csc<T: Scalar>(m: &CscMatrix<T>, x: &[T]) -> Result<Vec<T>, Dimens
     let row_idx = m.row_indices();
     let vals = m.values();
     for c in 0..m.cols() {
-        for k in col_ptr[c]..col_ptr[c + 1] {
-            let r = row_idx[k];
+        for k in col_ptr[c] as usize..col_ptr[c + 1] as usize {
+            let r = row_idx[k] as usize;
             let prev = out[r];
             out[r] = prev.add(vals[k].mul(x[c]));
         }
